chore(wizard-refactor): extract tools/suggest_welcome_message
Template-based suggestion generator. No pinning tests; dispatcher 28/28
holds the public contract green.
